+++ homework1/task1/ellipse.h
@@ -3,18 +3,18 @@
 class Ellipse
 {
     public:
-        int h, k, a, b;
+        double h, k, a, b;
 
     public:
-    Ellipse(int h = 0, int k = 0, int a = 0, int b = 0);
+    Ellipse(double h = 0, double k = 0, double a = 0, double b = 0);
     Ellipse(const Ellipse &e);
     ~Ellipse();
-    void normalize_axis(int &a, int &b);
-    int count_hyperparametr();
-    void set_all(int h_coord, int k_coord, int par_a, int par_b);
-    void get_all(int &h_coord, int &k_coord, int &par_a, int &par_b);
-    int* point_coordinates();
-    int* focus_coordinates();
+    void normalize_axis(double &a, double &b);
+    double count_hyperparametr();
+    void set_all(double h_coord, double k_coord, double par_a, double par_b);
+    void get_all(double &h_coord, double &k_coord, double &par_a, double &par_b);
+    double* point_coordinates();
+    double* focus_coordinates();
     double count_loc_hord();
     double count_excentritet();
     void print_data();
